Add LM optimiser
And switch to descent predict function for calculating loss
in all cases. Tweak settings to be reasonable for LM and Adam.
Note that LM trains faster but adding in forces slows things down. Adam
with small amount of forces in loss seems to outperform LM. Also, fix
write_scatter so that the force differences plotted are meaningful.
Remove weighting of conformers.